d tenía menos bits: bug

+++ libraries/signer/signer.h
@@ -11,5 +11,5 @@ class Signer
     Signer();
     ~Signer();
 
-    uint64_t sign(std::string message, uint32_t d, uint64_t n);
+    uint64_t sign(std::string message, uint64_t d, uint64_t n);
 };
